share realm model files with app extension
