+++ Realm/RealmRegexModel.swift
@@ -39,11 +39,3 @@ final class RealmRegexModel: Object {
 }
 
 extension RealmRegexModel: Identifiable { }
-
-extension RealmRegexModel {
-    public static func createNew() throws -> RealmRegexModel {
-        let model = ComponentsModel(components: [])
-        let componentsData = try JSONEncoder().encode(model.components)
-        return .init(id: UUID(), componentsData: componentsData)
-    }
-}
